+++ src/ally.js
@@ -7,7 +7,6 @@ import focusSource from './style/focus-source';
 import focusWithin from './style/focus-within';
 
 import maintainDisabled from './maintain/disabled';
-import maintainFocusTrapped from './maintain/focus-trapped';
 import maintainHidden from './maintain/hidden';
 
 import whenFocusable from './when/focusable';
@@ -50,7 +49,6 @@ export default {
 
   maintain: {
     disabled: maintainDisabled,
-    focusTrapped: maintainFocusTrapped,
     hidden: maintainHidden,
   },